web/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-phonenumber-field==8.0.0
phonenumbers==8.13.54
sqlparse==0.5.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.0